+++ Droid/MainActivity.cs
@@ -66,6 +66,7 @@ namespace PaketGlobal.Droid
             ZXing.Net.Mobile.Forms.Android.Platform.Init();
             Xamarin.FormsMaps.Init(this, bundle);
 			Xamarin.FormsGoogleMaps.Init(this, bundle, null); 
+            Stormlion.PhotoBrowser.Droid.Platform.Init(this);
 
             Countly.SharedInstance().Init(this, Config.CountlyServerURL, Config.CountlyAppKey).EnableCrashReporting();
             //Countly.SharedInstance().SetLoggingEnabled(true);